pyproject.toml
[project]
name = "sherrbyte"
version = "0.1.0"
description = "AI-powered hyper-personalized news platform for India"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "Proprietary" }

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",

    # Config & validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Observability
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.15.0",

    # Phase B — auth
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.9.0",
    "email-validator>=2.2.0",
    "slowapi>=0.1.9",

    # Phase C — RSS ingestion + dedup
    "feedparser>=6.0.11",
    "trafilatura>=1.12.0",
    "datasketch>=1.6.5",
    "langdetect>=1.0.9",
    "tenacity>=9.0.0",

    # Phase D — AI pipeline
    "google-genai>=0.3.0",
    "groq>=0.11.0",
    "redis>=5.1.0",
    "pgvector>=0.3.6",
    "sentence-transformers>=3.1.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.9",
    "mypy>=1.11.0",
    "types-redis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ---------------- Ruff ----------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RET", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------- Mypy ----------------
[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

# ---------------- Pytest ----------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra --strict-markers"

# ---------------- Coverage ----------------
[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
